Update standard_output in photoz.py
Modified standard_output in photoz.py to record the max-lnp photometric redshift even if the user requests final fits to be done at some other redshifts. 

If the user passes a specific set of "best" redshifts to use, EAZY forces the fit at these redshifts and records the fit information under "z_phot". The modification allows for new columns to be created for the max-lnp redshifts (provisionally called "z_pdf") so that this information is retained. If the user does not pass a specific set of redshifts, then EAZY defaults to fitting at max-lnp.